autogen_demo/requirements.txt
pyautogen>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.9.1
flake8>=6.1.0
mypy>=1.5.1